requirements.txt
pytest
pytest-cov
orderedset
flask
pycryptodome
mock
requests
